+++ src/main/java/kg/attractor/jobsearch_remake/service/ResumeService.java
@@ -53,7 +53,7 @@ public class ResumeService {
     @Transactional(readOnly = true)
     public List<ResumeDto> getByApplicant(Integer applicantId) {
         log.info("Получение резюме соискателя id: {}", applicantId);
-        return resumeRepository.findByApplicantId(applicantId).stream()
+        return resumeRepository.findByApplicantId(applicantId.longValue()).stream()
                 .map(this::toDto)
                 .toList();
     }
@@ -63,13 +63,14 @@ public class ResumeService {
         log.info("Получение резюме соискателя id: {}, страница: {}", applicantId, page);
         Pageable pageable = PageRequest.of(page, size,
                 Sort.by(Sort.Order.desc("updateTime")));
-        return resumeRepository.findByApplicantId(applicantId, pageable).map(this::toDto);
+        return resumeRepository.findByApplicantId(applicantId.longValue(), pageable)
+                .map(this::toDto);
     }
 
     @Transactional(readOnly = true)
     public ResumeDto getById(Integer id) {
         log.info("Получение резюме по id: {}", id);
-        return resumeRepository.findById(id)
+        return resumeRepository.findById(id.longValue())
                 .map(this::toDto)
                 .orElseThrow(() -> {
                     log.error("Резюме не найдено с id: {}", id);
@@ -90,15 +91,15 @@ public class ResumeService {
                 .updateTime(LocalDateTime.now())
                 .build();
         Resume saved = resumeRepository.save(r);
-        workExperienceInfoService.createForResume(saved.getId(), dto.getWorkExperienceInfos());
-        educationInfoService.createForResume(saved.getId(), dto.getEducationInfos());
-        contactInfoService.createForResume(saved.getId(), dto.getContactInfos());
+        workExperienceInfoService.createForResume(saved.getId().intValue(), dto.getWorkExperienceInfos());
+        educationInfoService.createForResume(saved.getId().intValue(), dto.getEducationInfos());
+        contactInfoService.createForResume(saved.getId().intValue(), dto.getContactInfos());
     }
 
     @Transactional
     public void update(Integer id, ResumeDto dto) {
         log.info("Обновление резюме id: {}", id);
-        Resume r = resumeRepository.findById(id)
+        Resume r = resumeRepository.findById(id.longValue())
                 .orElseThrow(() -> new NoSuchElementException("Резюме не найдено: " + id));
         r.setName(dto.getName());
         r.setCategoryId(dto.getCategoryId());
@@ -117,12 +118,12 @@ public class ResumeService {
         workExperienceInfoService.deleteByResumeId(id);
         educationInfoService.deleteByResumeId(id);
         contactInfoService.deleteByResumeId(id);
-        resumeRepository.deleteById(id);
+        resumeRepository.deleteById(id.longValue());
     }
 
     private ResumeDto toDto(Resume r) {
         return ResumeDto.builder()
-                .id(r.getId())
+                .id(r.getId().intValue())
                 .applicantId(r.getApplicantId())
                 .name(r.getName())
                 .categoryId(r.getCategoryId())
@@ -130,9 +131,9 @@ public class ResumeService {
                 .isActive(r.isActive())
                 .createdDate(r.getCreatedDate())
                 .updateTime(r.getUpdateTime())
-                .workExperienceInfos(workExperienceInfoService.getByResumeId(r.getId()))
-                .educationInfos(educationInfoService.getByResumeId(r.getId()))
-                .contactInfos(contactInfoService.getByResumeId(r.getId()))
+                .workExperienceInfos(workExperienceInfoService.getByResumeId(r.getId().intValue()))
+                .educationInfos(educationInfoService.getByResumeId(r.getId().intValue()))
+                .contactInfos(contactInfoService.getByResumeId(r.getId().intValue()))
                 .build();
     }
 }
